MeshViewer: split mesh rendering to single module
